+++ ui/src/main/kotlin/com/processout/sdk/ui/card/tokenization/CardTokenizationScreen.kt
@@ -21,7 +21,7 @@ import androidx.compose.ui.platform.rememberNestedScrollInteropConnection
 import androidx.compose.ui.res.colorResource
 import androidx.lifecycle.Lifecycle
 import com.processout.sdk.ui.card.tokenization.CardTokenizationEvent.*
-import com.processout.sdk.ui.card.tokenization.CardTokenizationSection.Item
+import com.processout.sdk.ui.card.tokenization.CardTokenizationViewModelState.Item
 import com.processout.sdk.ui.core.component.POActionsContainer
 import com.processout.sdk.ui.core.component.POHeader
 import com.processout.sdk.ui.core.component.POText
@@ -29,9 +29,8 @@ import com.processout.sdk.ui.core.component.field.PODropdownField
 import com.processout.sdk.ui.core.component.field.POField
 import com.processout.sdk.ui.core.component.field.POTextField
 import com.processout.sdk.ui.core.state.POActionState
+import com.processout.sdk.ui.core.state.POFieldState
 import com.processout.sdk.ui.core.state.POImmutableList
-import com.processout.sdk.ui.core.state.POMutableFieldState
-import com.processout.sdk.ui.core.state.POStableList
 import com.processout.sdk.ui.core.style.POAxis
 import com.processout.sdk.ui.core.theme.ProcessOutTheme
 import com.processout.sdk.ui.shared.composable.AnimatedImage
@@ -41,8 +40,7 @@ import com.processout.sdk.ui.shared.composable.rememberLifecycleEvent
 
 @Composable
 internal fun CardTokenizationScreen(
-    state: CardTokenizationState,
-    sections: POStableList<CardTokenizationSection>,
+    state: CardTokenizationViewModelState,
     onEvent: (CardTokenizationEvent) -> Unit,
     style: CardTokenizationScreen.Style = CardTokenizationScreen.style()
 ) {
@@ -83,7 +81,6 @@ internal fun CardTokenizationScreen(
         ) {
             Sections(
                 state = state,
-                sections = sections,
                 onEvent = onEvent,
                 style = style
             )
@@ -93,8 +90,7 @@ internal fun CardTokenizationScreen(
 
 @Composable
 private fun Sections(
-    state: CardTokenizationState,
-    sections: POStableList<CardTokenizationSection>,
+    state: CardTokenizationViewModelState,
     onEvent: (CardTokenizationEvent) -> Unit,
     style: CardTokenizationScreen.Style
 ) {
@@ -102,7 +98,7 @@ private fun Sections(
         LocalFocusManager.current.clearFocus(force = true)
     }
     val lifecycleEvent = rememberLifecycleEvent()
-    sections.elements.forEach { section ->
+    state.sections.elements.forEach { section ->
         section.title?.let {
             with(style.sectionTitle) {
                 POText(
@@ -112,7 +108,7 @@ private fun Sections(
                 )
             }
         }
-        section.items.forEach { item ->
+        section.items.elements.forEach { item ->
             Item(
                 item = item,
                 onEvent = onEvent,
@@ -161,7 +157,7 @@ private fun Item(
         is Item.Group -> Row(
             horizontalArrangement = Arrangement.spacedBy(ProcessOutTheme.spacing.small)
         ) {
-            item.items.forEach { groupItem ->
+            item.items.elements.forEach { groupItem ->
                 Item(
                     item = groupItem,
                     onEvent = onEvent,
@@ -178,7 +174,7 @@ private fun Item(
 
 @Composable
 private fun TextField(
-    state: POMutableFieldState,
+    state: POFieldState,
     onEvent: (CardTokenizationEvent) -> Unit,
     lifecycleEvent: Lifecycle.Event,
     focusedFieldId: String?,
@@ -229,7 +225,7 @@ private fun TextField(
 
 @Composable
 private fun DropdownField(
-    state: POMutableFieldState,
+    state: POFieldState,
     onEvent: (CardTokenizationEvent) -> Unit,
     fieldStyle: POField.Style,
     menuStyle: PODropdownField.MenuStyle,
@@ -246,7 +242,15 @@ private fun DropdownField(
             )
         },
         availableValues = state.availableValues ?: POImmutableList(emptyList()),
-        modifier = modifier,
+        modifier = modifier
+            .onFocusChanged {
+                onEvent(
+                    FieldFocusChanged(
+                        id = state.id,
+                        isFocused = it.isFocused
+                    )
+                )
+            },
         fieldStyle = fieldStyle,
         menuStyle = menuStyle,
         enabled = state.enabled,
